refactor: extract collections filtering
